No Need to Hard Code CC in gccLinux64.mak
There is no need to hard code the value of CC to "gcc" in gccLinux64.mak, as this value will default to `cc` and with a normal GNU Toolchain `cc` will be `gcc` anyway. By removing the explicit definition this also allows users to specify their own compiler (such as clang) with something like `CC=clang make -f gccLinux64.mak`.